backend/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
reportlab==4.2.0
pytest==8.2.0
httpx==0.27.0
